+++ src/xqt-kotlinx-lsp/commonMain/xqt/kotlinx/lsp/textDocument/Definition.kt
@@ -6,6 +6,7 @@ import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 import xqt.kotlinx.lsp.types.Location
 import xqt.kotlinx.lsp.types.Position
+import xqt.kotlinx.lsp.types.TextDocumentIdentifier
 import xqt.kotlinx.lsp.types.TextDocumentPositionParams
 import xqt.kotlinx.rpc.json.protocol.*
 import xqt.kotlinx.rpc.json.serialization.JsonSerialization
@@ -87,7 +88,9 @@ data class GoToResponse(
  * The goto definition request is sent from the client to the server to resolve the definition
  * location of a symbol at a given text document position.
  *
- * @since 1.0.0
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
+ * @since 2.0.0
  */
 fun TextDocumentRequest.definition(
     handler: TextDocumentPositionParams.() -> GoTo
@@ -102,11 +105,13 @@ fun TextDocumentRequest.definition(
  * The goto definition request is sent from the client to the server to resolve the definition
  * location of a symbol at a given text document position.
  *
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
  * @param params the request parameters
  * @param responseHandler the callback to process the response for the request
  * @return the ID of the request
  *
- * @since 1.0.0
+ * @since 2.0.0
  */
 fun TextDocumentJsonRpcServer.definition(
     params: TextDocumentPositionParams,
@@ -122,18 +127,23 @@ fun TextDocumentJsonRpcServer.definition(
  * The goto definition request is sent from the client to the server to resolve the definition
  * location of a symbol at a given text document position.
  *
- * @param uri the text document's URI
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
+ * @param textDocument the text document
  * @param position the position inside the text document
  * @param responseHandler the callback to process the response for the request
  * @return the ID of the request
  *
- * @since 1.0.0
+ * @since 2.0.0
  */
 fun TextDocumentJsonRpcServer.definition(
-    uri: String,
+    textDocument: TextDocumentIdentifier,
     position: Position,
     responseHandler: (TypedResponseObject<GoTo, JsonElement>.() -> Unit)? = null
 ): JsonIntOrString = definition(
-    params = TextDocumentPositionParams(uri = uri, position = position),
+    params = TextDocumentPositionParams(
+        textDocument = textDocument,
+        position = position
+    ),
     responseHandler = responseHandler
 )