+++ src/xqt-kotlinx-lsp/commonMain/xqt/kotlinx/lsp/textDocument/DocumentHighlight.kt
@@ -7,6 +7,7 @@ import kotlinx.serialization.json.JsonPrimitive
 import kotlinx.serialization.json.buildJsonObject
 import xqt.kotlinx.lsp.types.Position
 import xqt.kotlinx.lsp.types.Range
+import xqt.kotlinx.lsp.types.TextDocumentIdentifier
 import xqt.kotlinx.lsp.types.TextDocumentPositionParams
 import xqt.kotlinx.rpc.json.protocol.*
 import xqt.kotlinx.rpc.json.serialization.*
@@ -117,7 +118,9 @@ data class DocumentHighlightResponse(
  * The document highlight request is sent from the client to the server to resolve the document
  * highlights for a given text document position.
  *
- * @since 1.0.0
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
+ * @since 2.0.0
  */
 fun TextDocumentRequest.documentHighlight(
     handler: TextDocumentPositionParams.() -> DocumentHighlight
@@ -132,11 +135,13 @@ fun TextDocumentRequest.documentHighlight(
  * The document highlight request is sent from the client to the server to resolve the document
  * highlights for a given text document position.
  *
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
  * @param params the request parameters
  * @param responseHandler the callback to process the response for the request
  * @return the ID of the request
  *
- * @since 1.0.0
+ * @since 2.0.0
  */
 fun TextDocumentJsonRpcServer.documentHighlight(
     params: TextDocumentPositionParams,
@@ -152,18 +157,23 @@ fun TextDocumentJsonRpcServer.documentHighlight(
  * The document highlight request is sent from the client to the server to resolve the document
  * highlights for a given text document position.
  *
- * @param uri the text document's URI
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
+ * @param textDocument the text document
  * @param position the position inside the text document
  * @param responseHandler the callback to process the response for the request
  * @return the ID of the request
  *
- * @since 1.0.0
+ * @since 2.0.0
  */
 fun TextDocumentJsonRpcServer.documentHighlight(
-    uri: String,
+    textDocument: TextDocumentIdentifier,
     position: Position,
     responseHandler: (TypedResponseObject<DocumentHighlight?, JsonElement>.() -> Unit)? = null
 ): JsonIntOrString = documentHighlight(
-    params = TextDocumentPositionParams(uri = uri, position = position),
+    params = TextDocumentPositionParams(
+        textDocument = textDocument,
+        position = position
+    ),
     responseHandler = responseHandler
 )