+++ src/xqt-kotlinx-lsp/commonMain/xqt/kotlinx/lsp/textDocument/SignatureHelp.kt
@@ -6,6 +6,7 @@ import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.buildJsonObject
 import xqt.kotlinx.lsp.base.UInteger
 import xqt.kotlinx.lsp.types.Position
+import xqt.kotlinx.lsp.types.TextDocumentIdentifier
 import xqt.kotlinx.lsp.types.TextDocumentPositionParams
 import xqt.kotlinx.rpc.json.protocol.*
 import xqt.kotlinx.rpc.json.serialization.*
@@ -197,7 +198,9 @@ data class SignatureHelpResponse(
  * The signature help request is sent from the client to the server to request signature
  * information at a given cursor position.
  *
- * @since 1.0.0
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
+ * @since 2.0.0
  */
 fun TextDocumentRequest.signatureHelp(
     handler: TextDocumentPositionParams.() -> SignatureHelp
@@ -212,11 +215,13 @@ fun TextDocumentRequest.signatureHelp(
  * The signature help request is sent from the client to the server to request signature
  * information at a given cursor position.
  *
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
  * @param params the request parameters
  * @param responseHandler the callback to process the response for the request
  * @return the ID of the request
  *
- * @since 1.0.0
+ * @since 2.0.0
  */
 fun TextDocumentJsonRpcServer.signatureHelp(
     params: TextDocumentPositionParams,
@@ -232,18 +237,23 @@ fun TextDocumentJsonRpcServer.signatureHelp(
  * The signature help request is sent from the client to the server to request signature
  * information at a given cursor position.
  *
- * @param uri the text document's URI
+ * __NOTE:__ In LSP 1.x, the `textDocument` parameter was an inlined `uri` parameter.
+ *
+ * @param textDocument the text document
  * @param position the position inside the text document
  * @param responseHandler the callback to process the response for the request
  * @return the ID of the request
  *
- * @since 1.0.0
+ * @since 2.0.0
  */
 fun TextDocumentJsonRpcServer.signatureHelp(
-    uri: String,
+    textDocument: TextDocumentIdentifier,
     position: Position,
     responseHandler: (TypedResponseObject<SignatureHelp?, JsonElement>.() -> Unit)? = null
 ): JsonIntOrString = signatureHelp(
-    params = TextDocumentPositionParams(uri = uri, position = position),
+    params = TextDocumentPositionParams(
+        textDocument = textDocument,
+        position = position
+    ),
     responseHandler = responseHandler
 )